Improve handling of different types in input
Refactor `String` -> `Guid` -> `Object` handling into `Store` method, in
order to allow for methods that take an `Object` as a parameter to be
able to recieve a list of complex objects (when they are getting an
`object[,]` which contains `Guid` strings). This also removes some
duplication between `Get...<T>` methods.

Implement handling of trees (`List<List<T>>`) from `object[,]` inputs.